Backend/requirements.txt
# Editable install with no version control (4676-Sp2021-PS5==1.0.0)
-e /Users/crosenberg/Desktop/Spring 2024/GRAD CV/PS4
absl-py==2.1.0
adal==1.2.7
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
azure==4.0.0
azure-applicationinsights==0.1.1
azure-batch==4.1.3
azure-cognitiveservices-speech==1.37.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.30.1
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-datalake-store==0.0.53
azure-eventgrid==1.3.0
azure-graphrbac==0.40.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.1
azure-mgmt==4.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.50.0
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-cdn==3.1.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==4.6.2
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-containerservice==4.4.0
azure-mgmt-cosmosdb==0.4.1
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==1.0.0
azure-mgmt-devspaces==0.1.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==1.0.0
azure-mgmt-eventhub==2.6.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-iotcentral==0.1.0
azure-mgmt-iothub==0.5.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-logic==3.0.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementgroups==0.1.0
azure-mgmt-managementpartner==0.1.1
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.1
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-network==2.7.0
azure-mgmt-notificationhubs==2.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-rdbms==1.9.0
azure-mgmt-recoveryservices==0.3.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.2.1
azure-mgmt-resource==2.2.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.1.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.9.1
azure-mgmt-storage==2.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-nspkg==3.0.2
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.7
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
babel==2.17.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.6.3
cachetools==5.3.3
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
COMTool==3.4.1
contourpy==1.0.7
cryptography==42.0.5
cvzone==1.6.1
cycler==0.11.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.0
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.12.0
Flask==2.3.2
Flask-hCaptcha==0.6.0
flatbuffers==24.3.25
fonttools==4.39.4
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
google-ai-generativelanguage==0.6.15
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==20.1.0
h5py==3.10.0
httplib2==0.22.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.4
imageio==2.33.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.21.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isodate==0.6.1
itsdangerous==2.1.2
jax==0.4.25
jaxlib==0.4.25
jedi==0.19.1
Jinja2==3.1.2
joblib==1.2.0
jsonlines==3.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
keras==3.1.1
kiwisolver==1.4.4
lazy_loader==0.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.11
mistune==3.0.2
ml-dtypes==0.3.2
MouseInfo==0.1.3
mpmath==1.3.0
msal==1.28.0
msrest==0.7.1
msrestazure==0.6.4.post1
multidict==6.0.5
namex==0.0.7
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==6.4.12
numpy==2.2.4
oauthlib==3.2.2
openai==0.27.10
opencv-contrib-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==23.1
pandas==2.2.1
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.4
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.2.2
pluggy==1.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.11.2
pydantic_core==2.33.1
PyGetWindow==0.0.9
Pygments==2.17.2
PyJWT==2.8.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyparsing==3.0.9
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pyqtgraph==0.13.7
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
pyte==0.8.2
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.2.0
QtAwesome==1.3.1
qtconsole==5.6.0
QtPy==2.4.1
referencing==0.35.1
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
rubicon-objc==0.5.0
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.2
Send2Trash==1.8.3
six==1.16.0
sounddevice==0.4.6
soupsieve==2.6
SQLAlchemy==2.0.19
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.1.0
tifffile==2023.12.9
tinycss2==1.3.0
tokenizers==0.13.3
tomli==2.0.1
torch==1.13.0
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4.1
tqdm==4.66.2
traitlets==5.9.0
transformers==4.25.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.3.8
widgetsnbextension==4.0.13
wrapt==1.16.0
yarl==1.9.4
zipp==3.16.2
